Fix bug causing unstow to delete dest directory when unfolding
During and unstow when nearly all actions have been collected
_collect_folding_actions() is called. It find candidates for folding
i.e. when a directory contains symlinks to files that all share the same
parent directory.

It just so happens that the dest directory passed into unstow could also
meet this same criteria and would be deleted in error.

This change adds a special case special case to make sure this does not
happen.

+++ dploy/stowcmd.py
@@ -261,7 +261,8 @@ class UnStow(AbstractBaseStow):
                                            other_links):
                         self._fold(source_parent, parent)
 
-                elif other_links_parent_count == 0:
+                elif (other_links_parent_count == 0
+                      and not utils.is_same_file(parent, self.dest_input)):
                     self.actions.add(actions.RemoveDirectory(self.subcmd, parent))
 
     def _fold(self, source, dest):
